Fix and add more UT
